Moved lengthy `@core` typings to subfiles and barrel index
