+++ packages/core/package.json
@@ -4,7 +4,7 @@
   "description": "Jimp core",
   "main": "dist/index.js",
   "module": "es/index.js",
-  "types": "index.d.ts",
+  "types": "types/index.d.ts",
   "files": [
     "dist",
     "es",
