+++ packages/core/types/index.d.ts
@@ -1,0 +1,8 @@
+export * from './etc';
+export * from './functions';
+export * from './plugins';
+export * from './utils';
+import {Jimp} from './jimp';
+
+export { Jimp };
+export default Jimp;
