+++ packages/core/types/utils.d.ts
@@ -1,0 +1,60 @@
+import {
+  WellFormedPlugin,
+  IllformedPlugin,
+  JimpType,
+  JimpPlugin,
+} from './plugins';
+
+// This is required as providing type arrays gives a union of all the generic
+// types in the array rather than an intersection
+export type UnionToIntersection<U> =
+  (U extends any ? (k: U) => void : never) extends ((k: infer I) => void) ? I : never;
+
+// The values to be extracted from a WellFormedPlugin to put onto the Jimp instance
+export type WellFormedValues<T extends WellFormedPlugin> = T['class'] &
+  T['constants'];
+
+// Util type for the functions that deal with `@jimp/custom`
+export type FunctionRet<T> = Array<() => T>;
+
+/**
+ * This conditional cannot be flipped. TS assumes that Q is `WellFormed` even
+ * it does not have the `class` or `constant` props. As a result, it will end
+ * up `undefined`. Because we're always extending `IllformedPlugin` on the
+ * plugins, this should work fine
+ */
+export type GetPluginVal<Q> = Q extends IllformedPlugin
+  ? Q
+  : WellFormedValues<Q>;
+
+type GetTypeFuncArrValues<TypeFuncArr> =
+  // Given an array of types infer `Q` (Q should be the type value)
+  TypeFuncArr extends Array<() => infer Q>
+    ? // Get the well formed value to add to the union
+      WellFormedValues<Q>
+    : // This should never be reached
+      undefined;
+
+type GetPluginFuncArrValues<PluginFuncArr> =
+  PluginFuncArr extends Array<() => infer Q>
+  ? // Get the plugin value, may be ill-formed or well-formed
+    GetPluginVal<Q>
+  : // This should never be reached
+    undefined;
+
+/**
+ * A helper type to get the values to be intersected with `Jimp` to give
+ * the proper typing given an array of functions for plugins and types
+ */
+export type GetIntersectionAddons<
+  TypesFuncArr extends FunctionRet<JimpType>,
+  PluginFuncArr extends FunctionRet<JimpPlugin>
+> = UnionToIntersection<
+  GetTypeFuncArrValues<TypesFuncArr> & GetPluginFuncArrValues<PluginFuncArr>
+>;
+
+/**
+ * While this was added to TS 3.5, in order to support down to TS 2.8, we need
+ * to export this and use it in sub-packges that utilize it
+ */
+export type Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
